only new table id datatype to BIGINT

+++ server/migrations/20190913075446-create-counselor-bill-settings.js
@@ -20,7 +20,7 @@ module.exports = {
           type: Sequelize.STRING
         },
         counselor_id: {
-          type: Sequelize.BIGINT
+          type: Sequelize.INTEGER
           // references: {
           //   model: "counselors",
           //   key: "id"
@@ -29,7 +29,7 @@ module.exports = {
           // onDelete: "cascade"
         },
         currency_id: {
-          type: Sequelize.BIGINT
+          type: Sequelize.INTEGER
           // references: {
           //   model: "currencies",
           //   key: "id"
